UPDATED: code refactored

+++ src/main/java/com/bridgelabz/springaddressbookapp/service/AddressbookAppService.java
@@ -3,49 +3,66 @@ package com.bridgelabz.springaddressbookapp.service;
 
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.bridgelabz.springaddressbookapp.contactmodel.AddressbookModel;
 import com.bridgelabz.springaddressbookapp.dto.AddressbookDTO;
+import com.bridgelabz.springaddressbookapp.exceptions.AddressbookAppException;
 import com.bridgelabz.springaddressbookapp.repository.AddressbookRepository;
 
 
 @Service
-public class AddressbookAppService {
+public class AddressbookAppService implements IAddressbookAppService{
 	
 	@Autowired
 	AddressbookRepository repo;
 
-	public AddressbookModel  saveDataInDB(AddressbookDTO dto) {
-		AddressbookModel newContact = new AddressbookModel(dto);
-		repo.save(newContact);
-		return newContact;
-	}
-
-	public List<AddressbookModel> getListOfContacts() {
-		List<AddressbookModel> allDataInAddressbook = repo.findAll();
-		return allDataInAddressbook;
+	public AddressbookModel  saveAddress(AddressbookDTO dto) {
+		AddressbookModel address = new AddressbookModel(dto);
+		repo.save(address);
+		return address;
+		
 	}
 	
-	/*public AddressbookModel getContactById(Integer id) {
-		AddressbookModel newContact = repo.getById(id);
-		System.out.println(newContact);
-		return newContact;
-	}*/
+	public List<AddressbookModel> getListOfAddresses() {
+		List<AddressbookModel> list = repo.findAll();
+		if(list.isEmpty()) {
+			throw new AddressbookAppException("Addressbook is empty");
+		}
+		else {
+		return list;
+		}
+	}
 	
-	public Optional<AddressbookModel> getContactbyId(Integer id) {
-		return repo.findById(id);
+	public AddressbookModel getAddressbyId(Integer id) {
+		List<AddressbookModel> allAddresses = repo.findAll();
+		AddressbookModel address = allAddresses.stream()
+				.filter(addressBookData->addressBookData.getId()==id)
+				.findFirst()
+				.orElseThrow(()-> new AddressbookAppException("Employee Not found"));
+		return address;
 	}
 
 	public AddressbookModel updateDateById(Integer id, AddressbookDTO dto) {
-		AddressbookModel newContact = new AddressbookModel(id,dto);
-		repo.save(newContact);
-		return newContact;
+		Optional<AddressbookModel> address = repo.findById(id);
+		if(address.isPresent()) {
+			System.out.println("ID is valid");
+			AddressbookModel newAddress = new AddressbookModel(id,dto);
+			repo.save(newAddress);
+			return newAddress;
+		}
+		else {
+			throw new AddressbookAppException("Employee not found");
+		}
 	}
 
 	public void deleteContact(Integer id) {
-		repo.deleteById(id);
+		Optional<AddressbookModel> address = repo.findById(id);
+		if(address.isPresent()) {
+			repo.deleteById(id);
+		}
+		else {
+			 throw new AddressbookAppException("Invalid id..please input valid id");
+		}
 	}
 }
